Modify Home page and create/refractor components
